requirements.txt
Django==5.0.6
djangorestframework==3.15.2
drf-spectacular==0.27.2
django-filter==24.3
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
whitenoise==6.7.0
cloudinary==1.41.0
django-cloudinary-storage==0.3.0
dj-database-url==2.3.0
requests==2.32.3
gunicorn==22.0.0
